requirements.txt
scrapy
Pillow
dateparser
elasticsearch
scrapy-deltafetch
